test: TestFixture 분리
