+++ src/test/java/laddergame/fixture/BooleanGeneratorFixture.java
@@ -1,0 +1,20 @@
+package laddergame.fixture;
+
+import laddergame.domain.BooleanGenerator;
+
+import java.util.ArrayDeque;
+import java.util.Deque;
+import java.util.List;
+
+public abstract class BooleanGeneratorFixture {
+
+    public static final BooleanGenerator TEST_BOOLEAN_GENERATOR = new BooleanGenerator() {
+        Deque<Boolean> deque = new ArrayDeque<>(List.of(true, false));
+        @Override
+        public boolean generate() {
+            Boolean polled = deque.pollFirst();
+            deque.addLast(polled);
+            return polled;
+        }
+    };
+}
