+++ src/test/java/laddergame/fixture/HeightFixture.java
@@ -1,0 +1,10 @@
+package laddergame.fixture;
+
+import laddergame.domain.Height;
+
+public abstract class HeightFixture {
+    public static final Height HEIGHT_VALUE_1 = new Height(1);
+    public static final Height HEIGHT_VALUE_2 = new Height(2);
+    public static final Height HEIGHT_VALUE_3 = new Height(3);
+    public static final Height HEIGHT_VALUE_4 = new Height(4);
+}
