+++ src/test/java/laddergame/fixture/NameFixture.java
@@ -1,0 +1,9 @@
+package laddergame.fixture;
+
+import laddergame.domain.Name;
+
+public abstract class NameFixture {
+    public static final Name NAME_ROSIE = new Name("rosie");
+    public static final Name NAME_HYENA = new Name("hyena");
+    public static final Name NAME_JAYON = new Name("jayon");
+}
